refactor : 전반적 폴더 구조 수정
