+++ src/components/text/text.js
